requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
aiosqlite
pydantic
python-jose[cryptography]
passlib[bcrypt]
email-validator
pydantic_settings
greenlet
aiosqlite
alembic
jinja2
bcrypt==3.2.2
passlib
openai-whisper
python-multipart
pydub
fastapi-mail
datetime
pandas
streamlit
plotly
google-generativeai